pyproject.toml
[project]
name = "rag-knowledge-base"
version = "0.1.0"
description = "AIOps RAG Knowledge Base System for IT Operations and Security"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langchain-elasticsearch>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "elasticsearch>=8.11.0",
    "unstructured[all-docs]>=0.12.0",
    "pypdf>=4.0.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pillow>=10.2.0",
    "openai>=1.10.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "sqlalchemy>=2.0.0",
    "pdf2image>=1.16.0",
    "Pillow>=10.0.0",
    "pandas>=2.0.0",
    # 视觉模型依赖（LM Studio API）
    "openai>=1.10.0",
    # OCR 依赖
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.6.0",
    # 新方法依赖：Google Gemini + EasyOCR
    "google-generativeai>=0.3.0",
    "easyocr>=1.7.0",
    "opencv-python>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "web"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ["py39"]
